Add Assimp model loading assets

+++ src/Graphics/Geometry.cpp
@@ -12,7 +12,7 @@ Geometry::~Geometry()
 void Geometry::Create(const std::vector<D3D11_INPUT_ELEMENT_DESC>& layout, const ComPtr<ID3DBlob>& blob)
 {
 	_layout = layout;
-	HRESULT hr = Graphic::GetInstance()->GetDevice()->CreateInputLayout(_layout.data(), _layout.size(),
+	HRESULT hr = Graphic::GetInstance()->GetDevice()->CreateInputLayout(_layout.data(), static_cast<UINT>(_layout.size()),
 		blob->GetBufferPointer(), blob->GetBufferSize(), _inputLayout.GetAddressOf());
 	CHECK(hr);
 }
