+++ src/Graphics/IndexBuffer.cpp
@@ -11,7 +11,7 @@ IndexBuffer::~IndexBuffer()
 
 void IndexBuffer::Create(const std::vector<UINT>& indices)
 {
-	_count = indices.size();
+	_count = static_cast<UINT>(indices.size());
 	D3D11_BUFFER_DESC desc = {};
 	{
 		desc.ByteWidth = sizeof(UINT) * _count;
